support libibnetdisc caching no-overwrite flag
Hey Sasha,

This patch adds a flag to support a no-overwrite caching flag for
libibnetdisc.

Al

--
Albert Chu
chu11@llnl.gov
Computer Scientist
High Performance Systems Division
Lawrence Livermore National Laboratory

From: Albert Chu <chu11@llnl.gov>
Date: Mon, 19 Apr 2010 11:09:37 -0700
Subject: [PATCH] support libibnetdisc caching no-overwrite flag
Message-Id: <1271702059.17987.207.camel@auk31.llnl.gov>
Mime-Version: 1.0
Content-Transfer-Encoding: 7bit

+++ libibnetdisc/src/ibnetdisc_cache.c
@@ -888,10 +888,18 @@ int ibnd_cache_fabric(ibnd_fabric_t * fabric, const char *file,
 		return -1;
 	}
 
-	if (!stat(file, &statbuf)) {
-		if (unlink(file) < 0) {
-			IBND_DEBUG("error removing '%s': %s\n",
-				   file, strerror(errno));
+	if (!(flags & IBND_CACHE_FABRIC_FLAG_NO_OVERWRITE)) {
+		if (!stat(file, &statbuf)) {
+			if (unlink(file) < 0) {
+				IBND_DEBUG("error removing '%s': %s\n",
+					   file, strerror(errno));
+				return -1;
+			}
+		}
+	}
+	else {
+		if (!stat(file, &statbuf)) {
+			IBND_DEBUG("file '%s' already exists\n", file);
 			return -1;
 		}
 	}
